[core/config] global renaming for consistency
